+++ crates/rust-client/src/mock.rs
@@ -39,7 +39,7 @@ use crate::{
         generated::{
             merkle::MerklePath,
             note::NoteSyncRecord,
-            responses::{NullifierUpdate, SyncNoteResponse, SyncStateResponse},
+            responses::{SyncNoteResponse, SyncStateResponse},
         },
         NodeRpcClient, RpcError,
     },
@@ -164,16 +164,6 @@ impl MockRpcApi {
         // Collect notes that are in the next block
         let notes = self.get_notes_in_block(next_block_num).collect();
 
-        // Collect nullifiers from the next block
-        let nullifiers = next_block
-            .created_nullifiers()
-            .iter()
-            .map(|n| NullifierUpdate {
-                nullifier: Some(n.inner().into()),
-                block_num: next_block_num.as_u32(),
-            })
-            .collect();
-
         SyncStateResponse {
             chain_tip: self.get_chain_tip_block_num().as_u32(),
             block_header: Some(next_block.header().into()),
@@ -181,7 +171,6 @@ impl MockRpcApi {
             accounts: vec![],
             transactions: vec![],
             notes,
-            nullifiers,
         }
     }
 
@@ -303,6 +292,7 @@ impl NodeRpcClient for MockRpcApi {
     async fn check_nullifiers_by_prefix(
         &mut self,
         _prefix: &[u16],
+        _block_num: BlockNumber,
     ) -> Result<Vec<(miden_objects::note::Nullifier, u32)>, RpcError> {
         // Always return an empty list for now since it's only used when importing
         Ok(vec![])