+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * Add check for empty pay to id notes (#714).
 * [BREAKING] Refactored authentication out of the `Client` and added new separate authenticators (#718).
+* [BREAKING] `CheckNullifiersByPrefix` now uses a starting block number (#749).
 
 ## 0.7.0 (2025-01-28)